Wrist sim class created
